shill: rtnl_message: handle length underflow
RTA_NEXT() performs subtraction on the length argument, and it's
important that this subtraction can produce a negative result, so that
the next iteration of the loop can understand that we ran off the end
(i.e., in RTA_OK()).

Note that:
 * the callers for this are already using 'int' for length
 * rtattr uses 'unsigned short' (i.e., 16-bit) for length, so we should
   already be violating previous bounds checks if the *real* length
   can't actually fit in 'int'

I unwittingly introduced this error when refactoring the parsing code
here in https://crrev.com/c/1991965.

BUG=chromium:1045487, chromium:1045803
TEST=fuzzer

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2023290
Tested-by: Brian Norris <briannorris@chromium.org>

+++ shill/net/rtnl_message.cc
@@ -43,13 +43,15 @@ const char* TypeToString(RTNLMessage::Type type) {
   }
 }
 
-std::unique_ptr<RTNLAttrMap> ParseAttrs(struct rtattr* data, size_t len) {
+std::unique_ptr<RTNLAttrMap> ParseAttrs(struct rtattr* data, int len) {
   RTNLAttrMap attrs;
   ByteString attr_bytes(reinterpret_cast<const char*>(data), len);
 
   while (data && RTA_OK(data, len)) {
     attrs[data->rta_type] = ByteString(
         reinterpret_cast<unsigned char*>(RTA_DATA(data)), RTA_PAYLOAD(data));
+    // Note: RTA_NEXT() performs subtraction on 'len'. It's important that
+    // 'len' is a signed integer, so underflow works properly.
     data = RTA_NEXT(data, len);
   }
